migrations/models/4_20240228185724_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `log_log` RENAME TO `login_log`;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `login_log` RENAME TO `log_log`;"""
